receiving messages class
this class receives the messages and determines if it is spam
just instantiate the object and then call the Test method, which accepts Strings. 
I left the print statement in there. if you want the print removed, you can just delete it and get the result from the result variable
If you need an example, I can post one or I can show you in the next meeting